+++ src/app/layouts/index.ts
@@ -1,0 +1,1 @@
+export { MainLayout } from './mainLayout'
