+++ src/pages/homePage/index.tsx
@@ -1,15 +1,14 @@
 import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import MainLayout from 'shared/layouts/mainLayout'
-
-import { Button } from 'shared/ui/button'
+import { Button } from 'shared/components'
+import { MainLayout } from 'app/layouts'
 
 const HomePage: FC = () => {
   const { t } = useTranslation()
 
   return (
-    <MainLayout>
+    <MainLayout pageTitle="HomePage">
       <Button>{t('logout')}</Button>
     </MainLayout>
   )
