--- src/shared/layouts/index.ts
@@ -1,1 +1,0 @@
-export * from './mainLayout'
