+++ src/widgets/layouts/logoSection/index.tsx
