+++ src/widgets/layouts/styledComponents/index.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+import { Button, Col } from 'antd'
+
+export const StyledToggleSiderButton = styled(Button)`
+  font-size: 16px !important;
+  width: 64px !important;
+  height: 64px !important;
+  margin-right: 8px;
+`
+
+export const AuthBlockWrapper = styled(Col)`
+  margin-right: 6px;
+`
